requirements.txt
discord.py
python-dotenv
numpy
boto3